+++ pkg/analyzer/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.37.1 (Not yet published)
+* Added the getters `isDartCoreList`, `isDartCoreMap`, `isDartCoreSet`, and
+  `isDartCoreSymbol` to `DartType`.
+
 ## 0.37.0
 * Removed deprecated getter `DartType.isUndefined`.
 * Removed deprecated class `SdkLibrariesReader`.